requirements.txt
psycopg2-binary
mlflow
scikit-learn
pandas
pandas_datareader
pandas_ta
mplfinance
tensorflow==2.11.0
yfinance
seaborn
hyperopt
fastparquet
prefect
evidently
google-cloud-bigquery
google-cloud-storage
gcsfs
pandas-gbq




